Move models into folder, break into `core.py` and `link.py`
