+++ database_client/dynamic_query_constructor.py
@@ -15,13 +15,15 @@ from database_client.db_client_dataclasses import (
     WhereMapping,
 )
 from database_client.subquery_logic import SubqueryParameters
-from database_client.models.core import (
+from database_client.models.implementations.core import (
     Agency,
     DataSourceExpanded,
     DataRequestExpanded,
 )
-from database_client.models.helpers import convert_to_column_reference
-from database_client.models.table_reference import SQL_ALCHEMY_TABLE_REFERENCE
+from database_client.models.table_reference import (
+    SQL_ALCHEMY_TABLE_REFERENCE,
+    convert_to_column_reference,
+)
 from middleware.enums import RecordTypes, Relations
 from utilities.enums import RecordCategories
