+++ database_client/models/implementations/core.py
@@ -99,17 +99,6 @@ def get_iter_model_list_of_dict(instance, attr_name: str):
     ]
 
 
-class LinkAgencyDataSource(StandardBase):
-    __tablename__ = Relations.LINK_AGENCIES_DATA_SOURCES.value
-
-    data_source_id: Mapped[str] = mapped_column(
-        ForeignKey("public.data_sources.id"), primary_key=True
-    )
-    agency_id: Mapped[str] = mapped_column(
-        ForeignKey("public.agencies.id"), primary_key=True
-    )
-
-
 class Agency(StandardBase, CountMetadata):
     __tablename__ = Relations.AGENCIES.value
 
@@ -174,17 +163,6 @@ class AgencyExpanded(Agency):
     # Some attributes need to be overwritten by the attributes provided by locations_expanded
     state_iso = Column(String)
     county_name = Column(String)
-
-
-class LinkAgencyLocation(StandardBase):
-    __tablename__ = Relations.LINK_AGENCIES_LOCATIONS.value
-
-    location_id: Mapped[int] = mapped_column(
-        ForeignKey("public.locations.id"), primary_key=True
-    )
-    agency_id: Mapped[int] = mapped_column(
-        ForeignKey("public.agencies.id"), primary_key=True
-    )
 
 
 class TableCountLog(StandardBase, CreatedAtMixin):
@@ -477,23 +455,11 @@ class DataSourceArchiveInfo(Base):
     next_cached: Mapped[Optional[timestamp]]
 
 
-class LinkDataSourceDataRequest(StandardBase, DataSourceIDMixin):
-    __tablename__ = Relations.LINK_DATA_SOURCES_DATA_REQUESTS.value
-
-    request_id: Mapped[int] = mapped_column(ForeignKey("public.data_requests.id"))
-
-
 class DataRequestsGithubIssueInfo(StandardBase, DataRequestIDMixin):
     __tablename__ = Relations.DATA_REQUESTS_GITHUB_ISSUE_INFO.value
 
     github_issue_url: Mapped[str]
     github_issue_number: Mapped[int]
-
-
-class LinkUserFollowedLocation(
-    StandardBase, CountMetadata, CreatedAtMixin, UserIDMixin, LocationIDMixin
-):
-    __tablename__ = Relations.LINK_USER_FOLLOWED_LOCATION.value
 
 
 class RecordCategory(StandardBase):
@@ -597,10 +563,6 @@ class PendingUser(StandardBase, CreatedAtMixin):
     validation_token: Mapped[Optional[text]]
 
 
-class LinkLocationDataRequest(StandardBase, LocationIDMixin, DataRequestIDMixin):
-    __tablename__ = Relations.LINK_LOCATIONS_DATA_REQUESTS.value
-
-
 class DependentLocation(Base):
     __tablename__ = Relations.DEPENDENT_LOCATIONS.value
     __mapper_args__ = {"primary_key": ["parent_location_id", "dependent_location_id"]}
@@ -666,26 +628,6 @@ class RecentSearch(StandardBase, CreatedAtMixin, UserIDMixin, LocationIDMixin):
     __tablename__ = Relations.RECENT_SEARCHES.value
 
 
-class LinkRecentSearchRecordCategories(StandardBase):
-    __tablename__ = Relations.LINK_RECENT_SEARCH_RECORD_CATEGORIES.value
-
-    recent_search_id: Mapped[int] = mapped_column(
-        ForeignKey("public.recent_searches.id")
-    )
-    record_category_id: Mapped[int] = mapped_column(
-        ForeignKey("public.record_categories.id")
-    )
-
-
-class LinkRecentSearchRecordTypes(StandardBase):
-    __tablename__ = Relations.LINK_RECENT_SEARCH_RECORD_TYPES.value
-
-    recent_search_id: Mapped[int] = mapped_column(
-        ForeignKey("public.recent_searches.id")
-    )
-    record_type_id: Mapped[int] = mapped_column(ForeignKey("public.record_types.id"))
-
-
 class RecentSearchExpanded(StandardBase, CountMetadata, UserIDMixin, LocationIDMixin):
     __tablename__ = Relations.RECENT_SEARCHES_EXPANDED.value
 
@@ -709,7 +651,6 @@ class ChangeLog(StandardBase, CreatedAtMixin):
 class NotificationLog(StandardBase, CreatedAtMixin):
     __tablename__ = Relations.NOTIFICATION_LOG.value
 
-    id: Mapped[int] = mapped_column(primary_key=True, autoincrement=True)
     user_count: Mapped[int]
 
 
@@ -720,15 +661,3 @@ class DistinctSourceURL(Base):
     original_url: Mapped[str]
     rejection_note: Mapped[str]
     approval_status: Mapped[str]
-
-
-# region Views
-
-
-class LinkLocationDataSourceView(Base):
-    __tablename__ = Relations.LINK_LOCATIONS_DATA_SOURCES_VIEW.value
-    location_id = Column(Integer, primary_key=True)
-    data_source_id = Column(Integer, primary_key=True)
-
-
-# endregion
